refactor(turso-drizzle-adapter): src 配下をドライバ軸ディレクトリに整理
公開エントリをドライバ軸 (./serverless / ./database) で分けたが、内部の
ファイル本体は src/ 直下にフラットに並んでいて構造が不揃いだった。
ドライバ別のサブディレクトリ drivers/{serverless,database}/ に集約し、
ファイル名から冗長な接頭辞 (turso-, from-turso-) を取り除く。

root エントリは index.ts ではなく sqlite-run-result.ts を直接指す形にした。
中身は SqliteRunResult 型定義のみで、index.ts という集約役の名前と内容が
合わなかったため。コーディング規約「ファイル名と export 名を揃える」にも
沿う。

Phase 4 で sync-react-native を drivers/sync-react-native/ として追加すれば
対称構造を維持できる。
