+++ packages/turso-drizzle-adapter/package.json
@@ -4,9 +4,9 @@
 	"private": true,
 	"type": "module",
 	"exports": {
-		".": "./src/index.ts",
-		"./serverless": "./src/serverless.ts",
-		"./database": "./src/database.ts"
+		".": "./src/sqlite-run-result.ts",
+		"./serverless": "./src/drivers/serverless/index.ts",
+		"./database": "./src/drivers/database/index.ts"
 	},
 	"scripts": {
 		"lint": "biome check .",
