+++ packages/turso-drizzle-adapter/src/drivers/database/apply-migrations.ts
@@ -1,8 +1,8 @@
 import type { Database } from "@tursodatabase/database";
 import { migrate } from "drizzle-orm/sqlite-proxy/migrator";
-import { createDrizzleFromTursoDatabase } from "./create-drizzle-from-turso-database";
-import { createDatabaseExecutor } from "./sqlite-proxy/create-database-executor";
-import { proxyTransaction } from "./sqlite-proxy/proxy-transaction";
+import { createDatabaseExecutor } from "../../sqlite-proxy/create-database-executor";
+import { proxyTransaction } from "../../sqlite-proxy/proxy-transaction";
+import { createDrizzleFromTursoDatabase } from "./create-drizzle";
 
 export const applyMigrations = async (
 	database: Database,
