+++ packages/turso-drizzle-adapter/src/drivers/database/create-drizzle.ts
@@ -1,16 +1,16 @@
-import type { Client } from "@tursodatabase/serverless/compat";
+import type { Database } from "@tursodatabase/database";
 import { drizzle } from "drizzle-orm/sqlite-proxy";
-import { createServerlessExecutor } from "./sqlite-proxy/create-serverless-executor";
-import { proxyExecute } from "./sqlite-proxy/proxy-execute";
+import { createDatabaseExecutor } from "../../sqlite-proxy/create-database-executor";
+import { proxyExecute } from "../../sqlite-proxy/proxy-execute";
 
 // Drizzle 公式が @tursodatabase/* 用ドライバーを提供したら剥がす暫定アダプタ。現状スコープは sqlite-proxy 経由の橋渡しのみ（ADR-029）
-export const createDrizzleFromTursoServerless = <
+export const createDrizzleFromTursoDatabase = <
 	TSchema extends Record<string, unknown>,
 >(
-	client: Client,
+	database: Database,
 	schema: TSchema,
 ) => {
-	const executor = createServerlessExecutor(client);
+	const executor = createDatabaseExecutor(database);
 	return drizzle(
 		async (sql, params, method) => proxyExecute(executor, sql, params, method),
 		{ schema },
