+++ packages/turso-drizzle-adapter/src/drivers/database/create-handle.test.ts
@@ -1,6 +1,6 @@
 import type { Database } from "@tursodatabase/database";
 import { afterEach, beforeEach, describe, expect, test } from "vitest";
-import { createTursoDatabaseHandle } from "./create-turso-database-handle";
+import { createTursoDatabaseHandle } from "./create-handle";
 
 describe("createTursoDatabaseHandle", () => {
 	describe("インメモリDBへ接続したとき", () => {
