+++ packages/turso-drizzle-adapter/src/drivers/database/index.ts
@@ -1,4 +1,4 @@
 // `@tursodatabase/database`（ネイティブ .node バインディング）に依存するエントリ。Vercel Serverless 等の Web bundle に混入させないため、`./serverless` とは別パスで分離している
 export { applyMigrations } from "./apply-migrations";
-export { createDrizzleFromTursoDatabase } from "./create-drizzle-from-turso-database";
-export { createTursoDatabaseHandle } from "./create-turso-database-handle";
+export { createDrizzleFromTursoDatabase } from "./create-drizzle";
+export { createTursoDatabaseHandle } from "./create-handle";
