+++ packages/turso-drizzle-adapter/src/drivers/serverless/create-drizzle.ts
@@ -1,16 +1,16 @@
-import type { Database } from "@tursodatabase/database";
+import type { Client } from "@tursodatabase/serverless/compat";
 import { drizzle } from "drizzle-orm/sqlite-proxy";
-import { createDatabaseExecutor } from "./sqlite-proxy/create-database-executor";
-import { proxyExecute } from "./sqlite-proxy/proxy-execute";
+import { createServerlessExecutor } from "../../sqlite-proxy/create-serverless-executor";
+import { proxyExecute } from "../../sqlite-proxy/proxy-execute";
 
 // Drizzle 公式が @tursodatabase/* 用ドライバーを提供したら剥がす暫定アダプタ。現状スコープは sqlite-proxy 経由の橋渡しのみ（ADR-029）
-export const createDrizzleFromTursoDatabase = <
+export const createDrizzleFromTursoServerless = <
 	TSchema extends Record<string, unknown>,
 >(
-	database: Database,
+	client: Client,
 	schema: TSchema,
 ) => {
-	const executor = createDatabaseExecutor(database);
+	const executor = createServerlessExecutor(client);
 	return drizzle(
 		async (sql, params, method) => proxyExecute(executor, sql, params, method),
 		{ schema },
