+++ packages/turso-drizzle-adapter/src/drivers/serverless/index.ts
@@ -1,7 +1,7 @@
 // `@tursodatabase/serverless/compat`（HTTP プロトコル）に依存するエントリ。Vercel Serverless / Edge / iOS など、ネイティブバインディングを持ち込めない環境向け
-export { applyMigrationsToTursoServerless } from "./apply-migrations-to-turso-serverless";
-export { createDrizzleFromTursoServerless } from "./create-drizzle-from-turso-serverless";
+export { applyMigrationsToTursoServerless } from "./apply-migrations";
 export {
 	createTursoServerlessClient,
 	type TursoServerlessClientConfig,
-} from "./create-turso-serverless-client";
+} from "./create-client";
+export { createDrizzleFromTursoServerless } from "./create-drizzle";
